backend/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.29,<1.0
sqlmodel>=0.0.22,<0.0.25
pytest>=7.4,<9.0
passlib[bcrypt]>=1.7,<2.0
PyJWT>=2.8,<3.0
python-multipart>=0.0.7,<0.1
pymupdf>=1.24,<2.0
python-dotenv>=1.0,<2.0
openai>=1.40,<2.0
fpdf2>=2.7,<3.0
